requirements.txt
PyQt6>=6.0.0
numpy>=1.20
opencv-python>=4.5
Pillow>=8.0
rawpy>=0.17
tifffile>=2021.0
